restructure files into include and src
