--- comm_backend/CMakeLists.txt
@@ -1,1 +1,0 @@
-target_include_directories(reconverse PUBLIC .)
