+++ include/CMakeLists.txt
@@ -1,0 +1,1 @@
+target_include_directories(reconverse PUBLIC .)
