+++ src/convcore.C
