+++ src/converse_internal.h
