+++ test/CMakeLists.txt
@@ -1,1 +1,3 @@
 add_reconverse_executable(queue queue.C)
+# The test `queue` needs to access internal headers
+target_include_directories(queue PRIVATE ../src/)
